Enhance motion annotation files and update simulation documentation
- Expanded the _label_aliases.json file with additional label mappings for various motion transitions, improving the classification framework.
- Updated multiple motion JSON files to refine segment timings and labels, incorporating new transition labels such as tpol_to_m, m_to_tpol, and others for better accuracy in motion classification.
- Revised the main.tex document to reflect changes in the simulation model, including the introduction of transition clips and updated evaluation metrics.

These updates aim to improve the robustness of the fish school simulator and enhance the clarity of the associated documentation.

+++ src/sim/io.py
@@ -54,10 +54,10 @@ def load_trajectory_csv(path: Path) -> tuple[np.ndarray, np.ndarray]:
 def mmss_to_frame(mmss: str, fps: float) -> int:
     parts = mmss.strip().split(":")
     if len(parts) == 2:
-        m, s = int(parts[0]), int(parts[1])
+        m, s = int(parts[0]), float(parts[1])
         total = m * 60 + s
     elif len(parts) == 3:
-        h, m, s = int(parts[0]), int(parts[1]), int(parts[2])
+        h, m, s = int(parts[0]), int(parts[1]), float(parts[2])
         total = h * 3600 + m * 60 + s
     else:
         raise ValueError(f"Bad timestamp: {mmss!r}")
